usm: Remove redundant duplication of flushing events
We had a duplicated implementation between prebuilt and runtime builds for
flushing events form the kernel to the user mode.
Instead, we use a single implementation that is shared.

Furthermore, moving the core-logic of the flushing into an helper function
to allow future expansion to other flushing methods.
